refactor: replace home page with app in the project
