+++ src/components/pages/AboutUs/AboutUs.js
@@ -1,0 +1,9 @@
+import Header from '../../Header/Header';
+import React from 'react';
+const AboutUs = () => {
+    return (React.createElement("div", null,
+        React.createElement(Header, null),
+        React.createElement("h2", null, "About Us"),
+        React.createElement("p", null, "Information")));
+};
+export default AboutUs;
