+++ src/components/pages/AboutUs/AboutUs.tsx
@@ -1,0 +1,13 @@
+import Header from '../../Header/Header';
+import React from 'react';
+
+const AboutUs = () => {
+  return (
+    <div>
+      <Header />
+      <h2>About Us</h2>
+      <p>Information</p>
+    </div>
+  );
+}
+export default AboutUs;
