+++ src/main.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import './index.css';
-import App from './components/pages/Home/Home';
+import App from './components/App';
 ReactDOM.createRoot(document.getElementById("root")).render(React.createElement(React.StrictMode, null,
     React.createElement(BrowserRouter, null,
         React.createElement(App, null))));
