+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 
 import './index.css';
-import App from './components/pages/Home/Home';
+import App from './components/App';
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
